refactor APIRequest and signIn directory
